signet: add new -signet_enforcescript parameter
The enforcescript parameter can be used in signets where there is an alternative signer that
creates invalid blocks that can be optionally seen as valid by a node. The idea is to have
a way to simply test reorg scenarios by switching to the less restrictive enforcement, which
would periodically produce blocks on a seemingly valid chain, that were then replaced by a
longer main chain.